trim SwapContainer to point to either swap or pool, rearrange file structure to bring up to date
